requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
decorator==5.1.1
dj-database-url==2.0.0
Django==4.2.1
django-environ==0.10.0
django-redis==5.2.0
django-render-block==0.9.2
django-templated-email==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
executing==1.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
outcome==1.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.1.1
PySocks==1.7.1
pytz==2023.3
redis==4.5.5
requests==2.30.0
selenium==4.9.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==2.0.2
wcwidth==0.2.6
wsproto==1.2.0
zope.interface==6.0
